backend/requirements.txt
Django==5.2.2
djangorestframework==3.16.0
django-environ==0.12.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
pillow==11.2.1
django-templated-mail==1.1.1
